gpio: adnp: Make the number of pins configurable.
The code will automatically grow the register space in powers of two.
So in order to support 48 pins, the controller will actually have to
provide 8 registers for each of the DDR, PLR, IER, ISR and PTR sets even
if 6 would be enough.

Furthermore, interrupts are now cleared on read of the ISR registers.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>