gpio: adnp: Fix deadlock when setting output direction.
Setting the output direction involves writing the DDR and setting the
pin level. However, since setting the output value acquires a lock which
is already held by setting the direction, things will deadlack.

This is solved by introducing a non-locking variant of setting the pin
level and use it while under the lock used to write the DDR.