input: sx8634: Make button touch debounce configurable.
This will add a new parameter to configure the button touch debounce value.
To debounce the button-touch-event, the number of samples used to sense a
touch can be set in a range from 0x00 = 1 sample up to 0x03 = 4 samples.
This can be done via platform data or as module parameter. The module
parameter has the higher priority.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-by: Thierry Reding <thierry.reding@avionic-design.de>