tm6000: Add parameter to keep urb bufs allocated.
On systems where it cannot be assured that enough continous memory is available
all the time it can be very useful to only allocate the memory once when it is
needed the first time. Afterwards the initially allocated memory will be
reused, so it is ensured that the memory will stay available until the driver
is unloaded.

Signed-off-by: Julian Scheel <julian@jusst.de>