arm: tegra: Initialize lvds on plutux platform.
It seems that the l4t driver assumes that there is always a lvds display.
This leads to issues in the case were lvds is not setup. For instance we
can not change screen resolution because the frequences are not ajusted
properly.
To workaround this behavior we initialize the lvds display but force to
use hdmi as primary display.

Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>