arm: tegra: tec-ng: Add a clock setup workaround to allow using the AHUB
The AHUB block found on the tegra30 needs all its clients to be clocked,
otherwise the whole SoC freezes when accessing the AHUB registers.
This means that all the I2S blocks and the SPDIF block must be clocked
before the AHUB can be used. However the clock framework found in the
L4T kernel doesn't allow representing this dependency.

As a workaround the L4T u-boot pre-setups all these clocks, but such
hacks can't make it mainline u-boot. To allow our kernel to be booted
with a mainline u-boot forcibly start clocking all the I2S blocks and
the SPDIF block.

Change-Id: Ia632aae37e658ef49a17d4a85b96e77b4561d94d
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/241
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>