arm: tegra: tec-ng: Update defconfig.
Enable all actually required options for a stable tec-ng system. This includes
clock preinit, which is required to boot at all, as well as cpu frequency
scaling, to avoid overheating. Also for this to work thermal support is enabled.
Besides this some more features, which were previously untested are activated
now.

Change-Id: I6fa2f5193b9dd7cc0e1062264717e68b5774c829
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/456
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>