arm: tegra: tamonten: Add the nct1008 temperature sensor
The temperature sensor was only defined for the medcom-wide board,
however it is on the COM. So remove it from the medcom-wide board file
and add it back to the tamonten COM file.

Also enable the IRQ as the new driver bail out unless a valid IRQ is
given, or IRQ is explicitly disabled.

Change-Id: Ieb525d5fd7f7877fb2d99d64ddf795f6f7f1ee31
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/435
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>