media: tegra: There is 2 device named 'vi', rename one to 'vi01'
There is two devices named 'vi', one in the arch devices and one in the
host1x devices. This lead to a clash as both then need the same sysfs
node.

The host1x device could only be accessed with some hack, the arch one
is easy to get. So rename the host1x device to 'vi01' to avoid the
clash.

Change-Id: Ibe0414791ef7009bf5aafb44dd82f840d7b52005
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/466
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>