media: tegra: Fix all the stride computations
The stride computations all used soc_mbus_bytes_per_line() to compute
the stride of the output buffer. However this function is to compute
the line size of the data on the bus, it is not usable for planar
formats.

Also fix setting up the VI_VB0_BUFFER_STRIDE_FIRST register to give
the proper stride for luma as well as chroma.

Change-Id: I96965a28e671675b114c88d4b392ca01beeeb3e8
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/469
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>