staging: Backport zram driver from kernel 3.12.
zRAM increases performance by avoiding paging on disk and instead uses a
compressed block device in RAM in which paging takes place until it is
necessary to use the swap space on the hard disk drive. Since using RAM
is faster than using disks, zram allows Linux to make more use of RAM
when swapping/paging is required, especially on older computers with
less RAM installed.

Even though the cost of RAM hardware is relatively low, this feature
still offers advantages for netbooks and other lower-powered laptops,
virtualization and in the case of embedded devices, especially those
that use flash memories that have a limited lifespan, dependent on
write-use and thus wear out quickly when used as a swap device.

Also zRAM is used on Android and ChromeOS for a while now, see:
http://www.chromestory.com/2013/03/google-enabling-zram-for-chrome-os-by-default

Change-Id: Ic52139e12e9c78d175f178f1403efb2c29d72914
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/767
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>