asoc: tegra: Fix the CPU DAI binding on Tamonten COM
The binding from the codec to the CPU DAI is in fact hardcoded for the
T20 platforms. The T30 platforms would just patch the table according
to the machine type. However on Tamonten boards the machine type
doesn't imply the COM module type.

To fix this use some ifdef mechanism to directly populate the table
with the correct data.

Change-Id: Ibdfa3e68f9fbb7f1b2d1428f7f32a9f94576834b
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/519
Reviewed-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>