media: tegra: Wait for the DMA transfer end instead of vsync
The driver was waiting for the vsync event before returning the new
picture. However the picture is only available in memory once the DMA
transfer is done. This lead to massive output corruption as the data
was often not fully written before the userspace tried to read it.

Fix this bug by simply waiting for the DMA transfer end event instead
of the vsync event.

Change-Id: Ib2c6d21e79a22826a469c46b474b9e89122671a9
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/546
Reviewed-by: Julian Scheel <julian.scheel@avionic-design.de>