HACK: hid: Reset usb device when receiving to many invalid values.
Due to a mechanical construction issue, it is possible that the touch creates
events which do not result from a actual touch-event. These events then occur
in large amounts. To prevent this event flooding we reset the device when we
detect this behavior.

This workaround is only applied to one specific device.

Signed-off-by: Hendrik Lippek <hendrik.lippek@avionic-design.de>
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Change-Id: Ia669ed49c7a0241c54838556f5ac2d65711bbfdb
Reviewed-on: http://review.adnet.avionic-design.de/843
Reviewed-by: Julian Scheel <julian@jusst.de>
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>