arm: tegra: tamonten: Rework the allocation of board GPIOs and IRQs
The board GPIOs and IRQs are allocated sequentially after the COM
GPIOs and IRQs. The allocation must be static as the GPIOs and IRQs
are used in device declarations. This is a problem for our generic
board modules as the allocation can change with the config.

To solve this problem we add a new header, tamonten-board.h, that
list all the generic GPIOs and IRQs sources. In case one module is not
enable its count is just set to zero.

Change-Id: Ic46448252ba296ff970aa48af357a903f110ceac
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/699
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>