arm: tegra: tamonten: Fix the number of PMU IRQ on Tamonten (non-NG)
The number of IRQ used by the PMU must be properly defined, otherwise
board peripherals will try to use IRQ that are already in use.

The number of IRQ from the PMU wasn't clearly defines in the old code.
The base for the board device IRQ simply used an hard coded offset of
32, which worked because the PMU has less than 32 IRQs.

Change-Id: If40f3815bf3d72925346237630ae2601f605f0a5
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/698
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>