staging: zsmalloc: Ensure handle is never 0 on success
Commit 67296874eb1cc80317bf2a8fba22b494e21eb29b upstream.

zsmalloc encodes a handle using the pfn and an object index. On hardware
platforms with physical memory starting at 0x0 the pfn can be 0. This
causes the encoded handle to be 0 and is incorrectly interpreted as an
allocation failure.

This issue affects all current and future SoCs with physical memory
starting at 0x0. All MSM8974 SoCs which includes Google Nexus 5 devices are
affected.

To prevent this false error we ensure that the encoded handle will not be 0
when allocation succeeds.

Change-Id: I8538ba2a5f297cd7af36c0bc1160de098a8ddd23
Signed-off-by: Olav Haugan <ohaugan@codeaurora.org>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Reviewed-on: http://review.adnet.avionic-design.de/876
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>