arm: tegra: dvfs: Clamp the clock rate passed to tegra_dvfs_set_rate()
When a DVFS controlled clock is enabled, tegra_dvfs_set_rate() is
called with the current clock's rate. But when adjusting a clock
maximum rate, its current rate isn't changed. So it might happen that
tegra_dvfs_set_rate() is called with a rate outside of the clock's
allowed range.

Fix this problem by clamping the rate passed to tegra_dvfs_set_rate()
to the currently allowed range.

This fix allow the Tamonten NG to boot properly without requiring a
maximum EDP limit of 1300 mV.

Change-Id: If8e744a8ca935e95463aa47a12a80667ff4be0db
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/742
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>