input: Rework the SX8634 driver to make an MFD driver
Beside the touch controller the SX8634 also provides a GPIO controller
and PWMs. To properly support these we first rewrite the current input
driver to an MFD driver with a single sub device for the touch
controller.

Because of this the platform data struct needs to be changed a bit and
the platforms using it will need an update.

Change-Id: I314dea03b77963ce53a9557903f463e233f10103
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>