video: backlight: Add a backlight driver for the SX8634
This driver allow controlling backlights using the PWM outputs of the
SX8634. This feature of the SX8634 wasn't implemented using the PWM API
because the SX8634 doesn't allow to set the frequency, only the duty.
Moreover the SX8634 provides features, like the logarithmic curve, that
only make sense when driving LEDs.

Change-Id: If7a1aa8399dee52aa37efa95d6e72b7f65ff5db7
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/748
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>