mfd: tps6586x: add version detection
Use the VERSIONCRC to determine the exact device version. According to
the datasheet this register can be used as device identifier. The
identification is needed since some tps6586x regulators use a different
voltage table.

Based on a patch from Stefan Agner <stefan@agner.ch> that was extended
to also support the TPS658640.

Change-Id: I8d81b78027eaaf2aa16125db1f4e929fd136640c
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/777
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>