arm: tegra: tamonten: Disable temperature alert IRQ.
With or without IRQ, initialization works properly and the sensor is
accessible via sysfs interfaces. But if an IRQ is set, the sensor is not
available for libsensors in user-spaces, so tools using libsensors don't
get any data.

To workaround this, disable the IRQ for now, until we know the reason.

Change-Id: I0ff3418250b792d7a005d9f3abb8f978b636ebdd
Signed-off-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/772
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>