arm: tegra: tamonten: Use the ad7461 driver for the temperature sensor
The nct1008 driver is incomplete and doesn't implement the hwmon
interface. However the nct1008 is register compatible with the AD7461
which have an hwmon driver.

This also revert commit 56e18b6f2543547cd9c8900ea80a289b8ea823a4 that
just disabled the IRQ. This was in fact a missunderstanding as the
configuration tested with the IRQ disabled was in fact using the LM90
driver.

In current mainline the AD7461 driver has been dropped in favor of the
LM90. We still use the AD7461 driver here because the LM90 driver is
not configurable via platform data.

Change-Id: I609614d4b955410e2a0ef2f29030ca9e9d4cacf3
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Tested-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/781
Reviewed-by: Soeren Grunewald <soeren.grunewald@avionic-design.de>