gpio/adnp: Add an optional check for the MachOX CPLD
The ADNP is generally implemented on a MachOX CPLD. Attempting to use
the CPLD when it is not yet programmed put it in an unrecoverable state.

To avoid this add an optional check for this CPLD that calls a platform
function with the basic identifier from the CPLD (device id, usercode
and trace id). If the platform function returns an error the ADNP
driver isn't loaded. Additionally this function can be used to print
out information about the CPLD.

Change-Id: I44cf0c567eb9342ba5c58bbfe86c8d7235992a3a
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1017
Reviewed-by: Julian Scheel <julian@jusst.de>