arm: tegra: medcom-wide: Add deployed cpld platform id
The platform id of the cpld images deployed with medcom-wide gen1 (1427) ist
unfortunately not the one which it should be according to the documentation,
but an image originally made for Plutux. To make it even worse that images
platform id does not match the Plutux (1408) platform id specified in the
documentation.

Change-Id: I5ac0c3a5f99180aa8efee4619dd12e2c11848343
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/1294
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>