usb: gadget: core: add implementation of gadget bus
At gadget bus, the device (udc) and the driver (gadget dirver)
will try to match between each other, the automatically
binding is designed at this patch.

The rule of auto-binding is: if the udc is not in use and the
gadget driver is not used by any udcs, it is matched successfully,
and will be binded. Each gadget driver can only be binded at one
udc. For example, there are two udcs, "udc-0" and "udc-1", and
two gadget drivers, "g_mass_storage" and "g_serial", assumed the
udcs is added to bus first, then, the "udc-0" will be binded to
"g_mass_storage", and "udc-1" will be binded to "g_seiral".

Change-Id: I91d1358580fa2567bcb6828af6de2668ebf5d326
Signed-off-by: Peter Chen <peter.chen@freescale.com>
Reviewed-on: http://review.adnet.avionic-design.de/4226
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>