watchdog: tegra: Properly implement the watchdog on T20
The T20 watchdog is a bit particular, the timer must be setup for
continuous operation with overflow interrupt enabled. If the interrupt
flag is still set (from a previous interrupt) when the overflow
interrupt occurs the CPU is rested.

The current implementation for T20 would only clear the interrupt flag
in the interrupt handler and simply ignore the pings from the
userspace daemon.

Fix this by simply clearing the interrupt flag in the ping function.
We also need to set the interrupt trigger to raising edge, instead of
level high, to avoid an interrupt flood as the interrupt flag isn't
cleared in the handler.

Change-Id: If698ea82d0a612701d05ebb6caa2027d87ed7225
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1845
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>