hwmon: lm90: expose to thermal fw via DT nodes
This patch adds to lm90 temperature sensor the possibility
to expose itself as thermal zone device, registered on the
thermal framework.

The thermal zone is built only if a device tree node
describing a thermal zone for this sensor is present
inside the lm90 DT node. Otherwise, the driver behavior
will be the same.

Change-Id: I1b5d94a817ab3ea529c02e97f3df3d33fa838ba0
Signed-off-by: Wei Ni <wni@nvidia.com>
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3546
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>