ASoC: wm8903: Implement a proper debouce for the mic detect
The mic detect can bounce a lot during (un)plug events, on the other
hand the short detect has debounce already implemented in hardware.

To make things more robust split the reporting of mic and short
events, and move the mic reporting to a delayed work to properly
debounce it.

Change-Id: I467d1b0623ced55bfb8a2c5afb2cd55bdbb85d64
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1978
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>