ASoC: wm8903: Fix the mic detection with headset jacks
Headset jacks are backward compatible with typical 3 pins headphone
jacks, but they have a 4th pin for an optional microphone.

When plugging an headphone it short the mic, so we need to filter out
mic detect events when a short is also detected. To still allow using
the short detect for other purposes this behaviour is only enabled
when no event is bound to the short detection.

Change-Id: Ie0aa26b4b1a5ff5e2b8cb7cb0fadad3abb9b64f0
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/1979
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>