media: video: mt9v126: Cope with the sensor loosing its config
When running under a high ambient temperature the sensor sometimes
loose all its settings and need a hard reset to be usable again.
Once reseted the sensor must be fully configured again.

To cope with this move the sensor configuration to the state switching
function, that way the configuration can be set everytime it is needed.
This is then extended to hard reset the sensor if we detect that it
died.

Change-Id: I8558e6cdb551b37495652f1ce3b8f40370e02dd4
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2179
Reviewed-by: Julian Scheel <julian@jusst.de>