ARM: tegra: add basic DT support to the power detection code
The power detection code expect to be able to get all the regulator it
need. However with DT and defered probing we generally don't have the
regulators ready at this point.

To cope with this we replace the simple init call with a platform
driver to allow working with defered probe. No proper de-init has been
added, but the code has been fixed to work properly with multiple
calls failling with PROBE_DEFER at some point.

Change-Id: I5f717dd5640218122de9db4d82cf0c86e5f2856f
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/2633
Reviewed-by: Julian Scheel <julian@jusst.de>