ARM: tegra: meerkat: Remove the wakeup monitor device
This L4T specific module is not tied to any hardware; its main job is
broadcasting uevents on system suspend and resume. Optionally, it can
also gather network traffic statistics.

This looks like a dubious helper for Nvidia-specific userspace software
that we want to get rid of.  Unfortunately, the module does not support
devicetree.  While it could be deactivated as per CONFIG_TEGRA_WAKEUP_MONITOR,
that would require keeping references to such dubious and legacy code in
the meerkat driver, which we want to avoid wherever possible. So, just
remove the device from the meerkat driver, keeping it lean and clean.

Change-Id: If83485bef388f9e0c1f1e291d1981f7ac3b8ca83
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
[nikolaus.schulz@avionic-design.de: add rationale]
Reviewed-on: http://review.adnet.avionic-design.de/3093