ARM: tegra: meerkat: merge pinmux states
As an artifact from the Jetson pinmux, the meerkat pinmux defined two
extra pinmux states "drive" and "unused" in addition to the default.

Since these two states are unconditionally applied by the tegra pinctrl
driver as "user states" anyway, there is no point in maintaining this
distinction.  Moreover, it is very much possible for the pins in the
"unused" group to be used, which creates confusion.

Merging the two extra states into the default one resolves this, at the
price of a harmless warning by the pinctrl-tegra driver that it cannot
apply the two "user states".

Change-Id: Ic23c4f15b07565de67c6e7c7ee07011577eefe3e
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3240
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>