video: tegra: dc: of: Register dc port after initialisation
The dc port state table is used to keep track of the ports being in use. As
the driver initialisation might fail gracefully (i.e. probe deferral due to
unavailable regulators) it must not be set that early. Instead move it below
all dependency resolving, so that a later init after initial failure is
possible.

Change-Id: Id3b2b545bcee7b249046576153660b88f0aa0067
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/3256
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>