arm: tegra: Allow setting the PCIe lane owner from DT
The L4T kernel expects this parameter from the command line. However as this
describes the hardware config it makes much more sense to have it in the
device tree. So add a 'nvidia,lanes' node to the pcie-controller to configure
the lanes mode. This node should contain a value listed in
dt-bindings/pci/tegra-pcie.h. If present the DT config always takes precedence
over the command line parameter.

Change-Id: I51de79652a1a8cdc25604cae734108d681d920bd
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3318
Reviewed-by: Julian Scheel <julian@jusst.de>