pcie: tegra: Fix memory allocation from interrupt context
The tegra PCI driver does some GFP_KERNEL allocation and ioremap out
of the read and write function. However these are called from
interrupt context which prohibit such operation.

To workaround this problem we pre-allocate the needed memory for up to
16 bus and just fail if no free bus is available. 16 should be more
than enought for most embedded platform, but if the need arise it
would be trivial to make this configurable.

Change-Id: I4ab08154ca7f0be64d4eecc631d8240cf4f9a346
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3511
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>