asoc: tegra: Fix single channel output via I2S
By definition I2S transmits 2 frames unconditionally. Thus it is not valid to
double the bit time when one channel data is supplied. This just causes data
to be split between the two channels and generates a completely invalid
signal. By forcing a channel count of 2 for the clock computation the bit
length is valid all time and single channel streams are output on the first
i2s channel.

Change-Id: Ie0d6e20d986b904fb1478b2f68de8c24b9a0787c
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/3394
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>