ASoC: tegra: Fix setting the bit count for TDM mode
The tegra30_i2s_set_channel_bit_count() function implement a formula
given in the datasheet. However the datasheet is really unclear and it
seems that this function is only valid for 2 channels.

Instead the register should contains the number of clocks in one
frame, minus one. That only depend on the number of slots and the
number of bits per slot.

It is unclear if the NON_SYM_EN bit is ever needed in TDM mode, so we
don't set it, but leave a comment that this need to be cleared up.

Change-Id: Ie6a85d85bd1dfcef97110251ed42ef02543ced81
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3396
Reviewed-by: Julian Scheel <julian@jusst.de>