[media] soc_camera: Support sensor that use defered probe
Return the eventual error of soc_camera_probe() to
soc_camera_host_register() and abort in case of error.

This patch will break configs where some sensor is expected to fail.
If such config exists and need to be supported we could limit the new
behavior to -EPROBE_DEFER only.

Change-Id: Idad9bffd05f26fca6c8fca0ccc6b4d9b717b54f4
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3509
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>