arm: tegra: Use the maximum frequency when DVFS is defined using DT
The L4T kernel expects an extra command line parameter to allow DVFS
to use frequencies above the boot frequency. However the Meerkat SOM
usualy have a default BCT which is not at the maximum frequency and we
do want to use all the frequency defined in DT.

So if emc_max_dvfs is not set on the command line and the memory
controller is defined in DT allow to use the maximum frequency per
default. It is still possible to disable this behavior by explicitly
setting emc_max_dvfs=0 on the command line.

Change-Id: I25ade71de3fc12ec4c484dc81d36273249166820
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3830
Reviewed-by: Julian Scheel <julian@jusst.de>