ASoC: tegra: Fix the clock reconfiguration on Tobor
Because of limitations in the nvidia ASoC utils code all the external
clocks run at the same speed, reconfigure one change the rate of the
others. This lead to problems as the DAC was using different clock
speeds depending on the sample rate. Now that the DAC driver support
any clock rate we can use the same rate as for the ADC.

Change-Id: I8da43292cc016d84716e940bc7e000289ce14f99
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3886
Reviewed-by: Julian Scheel <julian@jusst.de>