usb: xhci: tegra: Cope with defered regulators
Since r21.4 the firmware loading was moved earlier, but that in turn
moved a large part of the init after the firmware load. However this
is out of the normal probe and deferal can't work properly.

To fix this we split getting and enabling the regulators in two
functions and run the get before loading the firmware.

Change-Id: Iaf566e001b86010788fca5a2006a896d85ae1c38
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3919
Reviewed-by: Julian Scheel <julian@jusst.de>