usb: ehci: tegra: Rework probing to support otg
Split parts of the probe and remove functions into the ehci_run and ehci_stop
hooks to allow dynamic starting and stopping of the ehci driver from the
tegra-otg driver. This is a necessary change for the following otg driver
rework, which previously recreated the whole ehci platform device for each
host-mode change.

Change-Id: I0b1851ef9acf9e0bb008248b968686e617d0cbc5
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/3975
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>