media: tegra: Request reset gpio for virtual sensor devices
Sensor devices may require toggling a reset gpio during powerup.  When
instantiating a virtual device, look for a reset gpio in the dt node
and request it, so that the device can be powered up.

Currently this is single-gpio only, but it can easily be extended to
multiple gpios when needed.

Change-Id: Iefc320c068f14300edfebbabaaf92c29f4084cf3
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4025
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>