ARM: tegra: usb: Use proper phy inst with device tree
Make use of a hack which was already included in the ehci driver, but not
used, that checks which phy interface to use use depending on the register
base address. This is needed for the phy and pmc driver to work and fixes a
memory corruption due to using pdev->id as phy interface index which is -1 in
case of device tree init. The same hack is done in ehci-tegra and tegra_udc
drivers now.

Change-Id: Iec69f669a127eeeff057e1ce99548be8518a9d24
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/4046
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>