usb: phy: tegra: Do disable phy frontend in otg mode
When the phy frontend is powered down while the USB controller is used in otg
mode the detection of current plugging state does not work reliable anymore.
Thus we skip disabling the phy frontend when the port has otg enabled to work
around this. Due to the bad design of the otg driver this causes the phy to be
not even powered down properly on driver unload, but this would not be easily
avoidable without doing a major rework of the whole driver construct.

Change-Id: I0e2cb5d98ecfbb59339d2b5f478a067ef3052407
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/4048
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>