media: tegra_camera: Reduce the data lane settle time
The data lane settle time should be proportional to the MIPI clock
rate, however in the VI it is programmed as number of cycles of the
CIL clock, which is fixed at 102MHz.

Idealy we should query the sensor about the clock speed used for the
current mode. As the V4L2 API currently doesn't allow this we have to
choose a value that accomodate most of the sensors we are currently
using. The current value is too high when the clock lane is too fast,
so reduce it to accomodate such sensors. The value is still high
enough for all the slow speed mode we are using.

Change-Id: Ifbf4a5c40eaf39df92ca065805b2424776d16ed2
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4121
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>