ALSA: ac97: Add option to set default record selection
The stk1160 only has line-in but there is no way to set the default record
selection in this kernel. In a current kernel we can set it in the reset
function but that ist not working here since the values gets overwritten.
We don't want to make it to complicated, so simply add a driver capability
that allows to select line-in as the default record selection.

Change-Id: I8d1804e21943858c0e9488b719d76c6d3d02172a
Signed-off-by: Dirk Leber <dirk.leber@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4149
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>