ARM: tegra12: Only power up the sensorhub on Jetson TK1
When we added the test to skip the Jetson sensorhub init we assumed
that the board ID would only be set on nvidia boards. However it turn
out that Parrot also set it in their config.

As this function should only run on Jetson TK1 we now check for the
Jetson TK1 board ID instead of just checking if it is set.

Change-Id: I5e4afd67fb67cffccc13126a5d4bf58b10f53669
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4507
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>