mmc: tegra: Make the polarity of the CD and WP GPIOs configurable
Some SD card slot don't use the typical polarity, extend the driver to
support this. For this two properties are added to the platform data
to select the polarity of each signal. To keep backward compatibility
they only need to be set for non standard polarities, so one force
high active while the other force low active.

Change-Id: If51fe2da258d670fd2e65fccb9b21b6a40861d84
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4417
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>