gpu: nvgpu: make frequency scaling non-optional
When kconfig creates .config from a tegra12_defconfig via make
oldconfig, it disables frequency scaling for the gk20a, which is
optional as per the defining Kconfig file, but enabled in the defconfig.
Until kconfig is fixed, work around that by making frequency scaling for
the gk20a non-optional, which is sufficient to have it preserved by make
oldconfig.

Change-Id: Iebefa4fe81959e1a2b91dcd214d5920fdc1107fb
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4385
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>