ARM: tegra: panel-generic: Support active_low gpio
Read the active low setting from device tree and honor it. Unfortunately
migration the driver to gpiod is non-trivial, because the gpios are not
specified under the devices node that the driver is triggered from. This is a
bit messy due to the way the tegra_dc driver deals with panels. Until we do a
bigger refactoring of this stick with gpio-legacy but read the polarity from
the dt.

Change-Id: I17f713c0f380533ae4f1bbd37283e9b3aee66ac2
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/4532
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>