ARM: tegra: meerkat: Move common usb properties from carrier boards to COM
A number of usb node properties have been duplicated in the devicetree
for every Meerkat carrier board, while in fact they describe the
hardware of the processor board.  Move those to the Meerkat dtsi.

This removes the properties nvidia,vbus-oc-map and nvidia,has-host from
the Meerkat carrier dtbs, the first being unused by the phy driver on
the tegra124, whereas the second should in fact be nvidia,has-hostpc,
but that is already defined in tegra124-soc.dtsi for all ports.

The situation with nvidia,hotplug is inconsistent in that for some
meerkat carrier boards it is set, while for others it's not; but since
that is still a sensitive area with bugs lurking in the drivers, this is
left as-is.

Change-Id: I6340d397cff44aa3afcf0956b03b87529df91df7
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4755
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>