ARM: tegra: tobor, tobor-aio: Supply sound devices with dummy reg properties
The reg property isn't used by the driver, but still the dt requires a
valid reg node for each device, so supply dummy ones for the tobor and
the tobor-aio.  This fixes a dtc warning as the tobor dt had sound reg
properties with invalid length; the tobor-aio had no reg property
defined for the sound devices.

Change-Id: Icc01aabf18c44b7182d5af7e6b9b7a288abaf15c
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4756
Reviewed-by: Julian Scheel <julian@jusst.de>