driver core: Add Kconfig switch for re-doing deferred probes early
Devices for which the driver has requested probe deferral are deferred
until their rescheduling is un-blocked in a late_initcall.  But if a
regulator probe is deferred, this means that its consumer devices are
deferred as well, which can delay the successful probe for these by
several seconds.

To speed up the boot process, add a Kconfig switch which enables
un-blocking the deferred probes earlier, with a subsys_initcall.

Change-Id: I39cfea2b7e8cdb3006eb0cba74a58a33f869d879
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5237
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>