i2c: tegra: Only shutdown the bus if all clients support it
Shuting down the bus when some of the clients are still running is
creating various problem. In particular with devices that use an
interrupt it prevent clearing the interrupt, but it can happend with
any thread that is using the device. For example the fan control
regularly read the temperature sensor.

To fix this issue check the clients on the bus and only shutdown the
bus if all clients support it.

Change-Id: I80eaadc6f0e9bad4eaf25fa95699a85920f9541b
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5213
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>