[media] tegra: vi2: Fix the deadlock when killing a capture
When an application that capture video is killed the video device FD
is closed without stopping the streaming first. This would dead lock
as we called vb2_fop_release() while holding the channel lock, however
the capture needs the lock to properly finish.

To fix this we first call vb2_fop_release() without the lock, then
take the lock to unset the input and decrement the use count.

Change-Id: I68e25bd5da52b0de607db33d7c74d85784948b7e
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5214
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>