[media] tegra: vi2: Remove unneeded enable of SOR powergate
The SOR powergate doesn't need to be enabled as it will
automatically get enabled with DSIA and DSIB.

Due to the way the unpowergate() function is designed,
the call to enable SOR would not check if it is already
enabled and thus would reset the SOR powergate.
This would cause the framebuffer on a DSI display to
break.

Change-Id: I2d1dead502d3b2bd5c9beeaeb95a31b6b288cb94
Signed-off-by: Marc Andre <marc.andre@netline.ch>
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
[alban.bedel@avionic-design.de: Also removed the corresponding
tegra_powergate_partition() call in the error path]
Reviewed-on: http://review.adnet.avionic-design.de/5259
Reviewed-by: Julian Scheel <julian@jusst.de>