video: tegra: fb: hdmi: Enable best videomode for current monitor specs
Instead of unconditionally trying to enable the first video mode from
the monitor specs, pick the best mode for the display that can be run by
the dc, honoring the preferred mode of the monitor if possible.

While at it, properly save the current mode when falling back to the
default mode.

Change-Id: Ie0c4137d74b7a964014d34fb107b58b3583eb47b
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5509
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>