[media] i2c: imx219: Fix the access to the driver private data
v4l2_i2c_subdev_init() calls i2c_set_clientdata() and set the client
data to the subdevice. This overwrite the driver own call to
i2c_set_clientdata(), so that i2c_get_clientdata() returns the
subdevice instead of the private struct. The driver was still working
because the subdevice happen to be the first member of the private
struct. Fix this by using container_of() to directly get the private
struct from the subdevice. While at it also remove the i2c_client
member from the private struct as it is not used.

Change-Id: If92bc9f880ef039538da323c4d4756a59013d3ce
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5590
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>