[media] i2c: uh2c: Fix the access to the driver private data
v4l2_i2c_subdev_init() calls i2c_set_clientdata() and set the client
data to the subdevice. This overwrite the driver own call to
i2c_set_clientdata(), so that i2c_get_clientdata() returns the
subdevice instead of the private struct. The driver was still working
because the subdevice happen to be the first member of the private
struct. Fix this by using container_of() to directly get the private
struct from the subdevice.

Change-Id: I901e8517d7846690acc81ced7bf271b04c1a68bd
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5591
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>