[media] tegra: vi2: Check if sensors are defined for the VI
After parsing the DT for sensors, print an error message and abort if
no sensor was defined. This give the user a better message as up to
now it would only print "Failed to register async notifier" which
isn't very helpful. Moreover it prevent going through the whole init
for nothing as we already know that there is no sensor.

Change-Id: I9038b46951a54fd11db00b7a5751935187e47f9d
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5594
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>