video: tegra: dc: Derive reserved framebuffer memory sizes from DT
Rather than hard-coding the amount of framebuffer memory that is
reserved for the display controllers during early boot, read from the DT
which dc are enabled, and assign them memory depending on the display
type that's connected.

The latter means that for a framebuffer console on an external display,
the fb memory should be large enough for displays with 4K resolution.
For internal displays, memory is reserved for displays up to Full HD
resolution.

Change-Id: I21f8101b15c60b318653d470b6db061daaee82d4
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5595
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>